refactor: group Flame map controls under controls submodule (#3878)
Move GameMapControls, GameMapCornerControls, and GameMapEmpireLeftRail
into features/game/flame/controls/ with a public barrel export. Update
imports, widgetbook catalog, and SPEC path references.

Refs #3878
