+++ SPEC/program/e2e-integration-tests.md
@@ -31,7 +31,7 @@
 | `ctE2eLastPanelSnapshot` | Province overlay: updated while open with valid `selectedTileKey`; cleared when closed. Separate **civilian / naval / production** snapshot updaters in `ct_e2e_last_panel_snapshot.dart` for those panels. |
 | Post–next-turn naval snapshot | When `CT_E2E=true`, `refreshCtE2eNavalPanelSnapshotAfterTurnIfEnabled` (in `ct_e2e_turn_snapshot_refresh.dart`, called from `applyTurnResolutionResult`) repopulates `ctE2eNavalPanelSnapshot` from current game providers after each resolved turn so fleet E2E loops can skip `openNavalPanel` when world state already shows NW arrival. The naval bottom sheet no longer clears this snapshot on close. |
 
-**Code:** `app/lib/config/ct_e2e.dart`, `app/lib/config/ct_e2e_last_panel_snapshot.dart`, `app/lib/features/game/flame/game_screen_shared.dart` (next-turn key), `app/lib/features/game/flame/game_map_controls.dart` (region tab keys), `app/lib/features/game/widgets/move_fleet_dialog.dart` (move dialog scroll root), `app/lib/features/game/widgets/fleet_expansion_tile.dart` + `app/lib/features/game/widgets/units/shared/units_entity_action_row.dart` (fleet Move action key).
+**Code:** `app/lib/config/ct_e2e.dart`, `app/lib/config/ct_e2e_last_panel_snapshot.dart`, `app/lib/features/game/flame/game_screen_shared.dart` (next-turn key), `app/lib/features/game/flame/controls/game_map_controls.dart` (region tab keys), `app/lib/features/game/widgets/move_fleet_dialog.dart` (move dialog scroll root), `app/lib/features/game/widgets/fleet_expansion_tile.dart` + `app/lib/features/game/widgets/units/shared/units_entity_action_row.dart` (fleet Move action key).
 
 **Game-start intro:** `e2eAdvanceGameStartIntroUntilDismissed` / `e2eGameStartIntroBlocksUi` in `e2e_test_shared.dart` advance all yarn lines and the final choice (`Continue` then `I shall.`) before panel opens. Bootstrap and `e2eDismissTransientUi` call this so map HUD arrival does not leave the intro shell blocking empire-rail taps.
