+++ SPEC/ui/player-turn-event-feed.md
@@ -9,7 +9,7 @@
 
 - Show significant outcomes relevant to the local human player as short "something happened!" lines.
 - Wide and narrow layouts: hidden-by-default feed panel over the map.
-- **Tab-bar row layout (LTR; mockup `.tabbar`, issue #2861 M1):** The newspaper toggle lives in the trailing slot of **[GameTabBar](../../app/lib/features/game/widgets/game_tab_bar.dart)** (inside [GameMapControls](../../app/lib/features/game/flame/game_map_controls.dart)) — not as a sibling overlay on the map `Stack`. The region tabs (`Old World`, `New World`) are **start-aligned** (left in LTR) inside an `Expanded` horizontal scroll view so they shrink without overflow on narrow viewports. A **flex spacer** (the `Expanded` region-tab area) separates the tabs from the trailing indicator group. The **treasury**, **cargo hold**, and **news toggle** are **end-aligned** on the trailing (right) edge in that order (treasury → cargo → news toggle), matching the mockup `.tabbar-spacer { flex:1 }` + trailing `.treasury` / `.cargo-hold` / `.news-toggle`. The toggle keeps a **4 dp leading margin** (`GameTabBar.clusterTrailingGap`) from the cargo hold indicator. This keeps the control out of the **320 logical px** wide province / sea zone side column on wide layouts so it does not cover the province overlay close affordance. **RTL:** out of scope for this change; layout is LTR-only.
+- **Tab-bar row layout (LTR; mockup `.tabbar`, issue #2861 M1):** The newspaper toggle lives in the trailing slot of **[GameTabBar](../../app/lib/features/game/widgets/game_tab_bar.dart)** (inside [GameMapControls](../../app/lib/features/game/flame/controls/game_map_controls.dart)) — not as a sibling overlay on the map `Stack`. The region tabs (`Old World`, `New World`) are **start-aligned** (left in LTR) inside an `Expanded` horizontal scroll view so they shrink without overflow on narrow viewports. A **flex spacer** (the `Expanded` region-tab area) separates the tabs from the trailing indicator group. The **treasury**, **cargo hold**, and **news toggle** are **end-aligned** on the trailing (right) edge in that order (treasury → cargo → news toggle), matching the mockup `.tabbar-spacer { flex:1 }` + trailing `.treasury` / `.cargo-hold` / `.news-toggle`. The toggle keeps a **4 dp leading margin** (`GameTabBar.clusterTrailingGap`) from the cargo hold indicator. This keeps the control out of the **320 logical px** wide province / sea zone side column on wide layouts so it does not cover the province overlay close affordance. **RTL:** out of scope for this change; layout is LTR-only.
 - **Floating feed card (wide):** When the province detail side panel is open (`mapProvincePanelProvider.overlayOpen` on wide), map-stack overlays that must stay in the map column use the same horizontal clearance as the region minimap: `Positioned.right = gameMapWideOverlayRightInset(provincePanelOpen: …)` from `game_screen_shared.dart` — i.e. **`kGameMapWideStackRightGutter` (8)** plus **`kGameMapWideProvinceSidePanelWidth` (320)** when the panel is open, else **8** only. The feed card uses this inset so **feed + open panel** does not overlap the panel column. When the panel is closed, behavior matches the prior **8** gutter. Cross-ref: `SPEC/ui/in-game-shell-narrow.md` (wide side panel width), `SPEC/ui/province-sea-zone-detail-overlay.md` (close control).
 - **Narrow:** Same toggle in **GameMapControls** (not over the map). The floating feed card remains `Positioned` on the map stack with the existing narrow inset; it must not obscure the bottom sheet’s primary chrome (including close). If both sheet and feed are open, verify stacking; relocate or inset only if a gap remains.
 - News icon button shows a badge with the current feed-entry count.
