+++ app/lib/features/game/flame/controls/controls.dart
@@ -1,0 +1,3 @@
+export 'game_map_controls.dart';
+export 'game_map_corner_controls.dart';
+export 'game_map_empire_left_rail.dart';
