+++ app/lib/features/game/flame/game_region_minimap.dart
@@ -348,7 +348,7 @@ class _MinimapZoomControlsState extends State<_MinimapZoomControls> {
 ///
 /// Mirrors mockup `.minimap-toggle`
 /// (`SPEC/ui/mockups/GAME10001-game-screen.html`) and the
-/// [GameMapCornerControls](game_map_corner_controls.dart) chrome family
+/// [GameMapCornerControls](../controls/game_map_corner_controls.dart) chrome family
 /// per `SPEC/ui/empire-overview.md` § Region minimap chrome
 /// (dark editorial-monocle): a flat `--bg-deep` surface with a 1 px
 /// `--border` outline, centered glyph tinted via `ColorFiltered(srcIn)`
@@ -372,7 +372,7 @@ class _MinimapToggleButton extends StatefulWidget {
   static const double iconSize = 20;
 
   /// Hover/press transition duration. Matches
-  /// [GameMapCornerControls](game_map_corner_controls.dart) so the row
+  /// [GameMapCornerControls](../controls/game_map_corner_controls.dart) so the row
   /// of map chrome reads as one editorial-monocle family.
   static const Duration animationDuration = Duration(milliseconds: 120);
   static const Curve animationCurve = Curves.easeOut;
