+++ app/lib/features/game/flame/map_state/game_map_area.dart
@@ -42,9 +42,7 @@ import '../../../../config/ui_screen_ids.dart';
 import '../game_screen_shared.dart';
 import '../game_side_menu.dart';
 import '../map_area/map_area.dart' show GameMapAreaBackground;
-import '../game_map_controls.dart';
-import '../game_map_corner_controls.dart';
-import '../game_map_empire_left_rail.dart';
+import '../controls/controls.dart';
 import '../game_map_canvas_stack.dart';
 import '../game_region_minimap.dart';
 import '../game_map_narrow_detail_overlay.dart';
