+++ app/test/players_bar_toggle_test.dart
@@ -1,4 +1,4 @@
-import 'package:colonizethis_app/features/game/flame/game_map_controls.dart';
+import 'package:colonizethis_app/features/game/flame/controls/controls.dart';
 import 'package:colonizethis_app/features/game/flame/game_screen_shared.dart'
     show kPlayersBarToggleButtonKey, kPlayerTurnFeedToggleButtonKey;
 import 'package:colonizethis_app/l10n/l10n.dart';
